requirements.txt
# local package
-e .

# external requirements
Sphinx
coverage
flake8
auto_mix_prep==0.2.0
click==8.1.3
matplotlib==3.6.3
numpy==1.24.1
python-dotenv==0.21.0
torch==1.13.1
torch_geometric==2.2.0
